+++ encog-core-cs/ML/Data/Market/MarketPoint.cs
@@ -43,7 +43,7 @@ namespace Encog.ML.Data.Market
         /// <param name="when">When is this data from.</param>
         /// <param name="size">What is the size of the data.</param>
         public MarketPoint(DateTime when, int size)
-            : base(size)
+            : base((UInt64)size)
         {
             _when = when;
         }
